# backend/swagger/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/users/avatar:
    delete:
      parameters:
      - description: Username
        in: query
        name: username
        required: true
        type: string
      responses:
        "200":
          description: Avatar deleted
      summary: Delete a user's avatar
      tags:
      - Users
    get:
      parameters:
      - description: Username
        in: query
        name: username
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
        "304":
          description: Avatar unchanged
        "404":
          description: No avatar
      summary: Get a user's avatar
      tags:
      - Users
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Username
        in: query
        name: username
        required: true
        type: string
      - description: JPEG, PNG, or WebP image
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload or replace a user's avatar
      tags:
      - Users
...
swagger: "2.0"